pyproject.toml
[project]
name = "slurm-executor"
version = "0.1.0"
description = "A library for transparent execution of heavy Python jobs on SLURM!"
readme = "README.md"
authors = [
    { name = "Stanisław Markowski", email = "stanislaw.markowski@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "cloudpickle>=3.1.1",
    "fabric>=3.2.2",
    "jinja2>=3.1.6",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
    "numpy>=2.3.4",
    "ruff>=0.14.0",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "Q"]
extend-select = ["E", "F"]